requirements.txt
slack-bolt
slack-sdk
google-generativeai
requests
rembg
Pillow
sqlalchemy
fastapi
uvicorn
jinja2
python-dotenv
aiofiles
uuid
datetime
pathlib
threading
asyncio
